+++ js/const.js
@@ -1,0 +1,42 @@
+const PHOTOS_MIN = 1;
+const PHOTOS_MAX = 25;
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+const COMMENTS_MIN = 0;
+const COMMENTS_MAX = 30;
+const AVATARS_MIN = 1;
+const AVATARS_MAX = 6;
+
+const DESCRIPTIONS = [
+  'Не откладывай на завтра, откладывай сразу навсегда',
+  'Однажды мне было весело, но это было ужасно',
+  'Худых людей похитить проще',
+  'Раньше я думала, что я нерешительная, но теперь я в этом не уверена',
+  'Мне нужны шестимесячные каникулы два раза в год',
+  'Нынешней весной зима особенно удалась'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = ['Василиса', 'Мира', 'Лёва', 'Ева', 'Дарина', 'Марк', 'Артём', 'Варя', 'Лёша', 'Лея', 'Арина', 'Мия', 'Платон', 'Доминик', 'Эвита'];
+
+export {
+  PHOTOS_MIN,
+  PHOTOS_MAX,
+  LIKES_MIN,
+  LIKES_MAX,
+  COMMENTS_MIN,
+  COMMENTS_MAX,
+  AVATARS_MIN,
+  AVATARS_MAX,
+  DESCRIPTIONS,
+  MESSAGES,
+  NAMES,
+};
