+++ js/util.js
@@ -1,0 +1,37 @@
+const getSequentNumber = () => {
+  let lastNumber = 0;
+  return function () {
+    lastNumber++;
+    return lastNumber;
+  };
+};
+const getRandomNumber = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+};
+
+const getRandomElement = (elements) =>
+  elements[getRandomNumber(elements.length - 1, 0)];
+
+function getRandomUniqueNumber (min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomNumber(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomNumber(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+export {
+  getSequentNumber,
+  getRandomNumber,
+  getRandomElement,
+  getRandomUniqueNumber,
+};
